+++ Tests/unit/test_materials.py
@@ -111,3 +111,36 @@ def test_non_matching_properties():
     my_mats = Materials([mat_1, mat_2])
     with pytest.raises(ValueError, match=r"rho is not defined for all materials"):
         my_mats.check_consistency()
+
+
+def test_check_borders():
+    materials = [
+        Material(id=1, D_0=None, E_D=None, borders=[0.5, 0.7]),
+        Material(id=2, D_0=None, E_D=None, borders=[0, 0.5]),
+            ]
+    size = 0.7
+    assert Materials(materials).check_borders(size) is True
+
+    with pytest.raises(ValueError, match=r'zero'):
+        size = 0.7
+        materials = [
+            Material(id=1, D_0=None, E_D=None, borders=[0.5, 0.7]),
+            Material(id=1, D_0=None, E_D=None, borders=[0.2, 0.5]),
+            ]
+        Materials(materials).check_borders(size)
+
+    with pytest.raises(ValueError, match=r'each other'):
+        materials = [
+            Material(id=1, D_0=None, E_D=None, borders=[0.5, 1]),
+            Material(id=1, D_0=None, E_D=None, borders=[0, 0.6]),
+            Material(id=1, D_0=None, E_D=None, borders=[0.6, 1]),
+            ]
+        size = 1
+        Materials(materials).check_borders(size)
+
+    with pytest.raises(ValueError, match=r'size'):
+        materials = [
+            Material(id=1, D_0=None, E_D=None, borders=[0, 1]),
+        ]
+        size = 3
+        Materials(materials).check_borders(size)
